refactor: ログ関連を shared/utils/log.ts に集約
